refactor: create migrator/ dir
