+++ great_expectations/__init__.py
@@ -3,7 +3,7 @@ from ._version import get_versions  # isort:skip
 
 __version__ = get_versions()["version"]  # isort:skip
 
-from great_expectations.data_context.cloud_migrator import CloudMigrator
+from great_expectations.data_context.migrator.cloud_migrator import CloudMigrator
 
 del get_versions  # isort:skip
